[ADD] sale_suggested_price: Split module out of beesdoo_product
The new module more or less works after this commit. beesdoo_product has
not been tested, and is probably in a bad way from this invasive
surgery. That's OK---it will recover from this.
