+++ beesdoo_product/__manifest__.py
@@ -33,8 +33,6 @@
         "views/purchase_views.xml",
         "views/sale_views.xml",
         "views/assets.xml",
-        "views/res_config_settings.xml",
-        "wizard/views/adapt_sales_price_wizard_view.xml",
         "wizard/views/label_printing_utils.xml",
         "security/ir.model.access.csv",
     ],
