+++ beesdoo_product/views/beesdoo_product.xml
@@ -14,13 +14,6 @@
                     attrs="{'invisible' : [('barcode','!=',False)]}"
                 />
             </field>
-            <field name="list_price" position="after">
-                <field
-                    name="suggested_price"
-                    widget='monetary'
-                    options="{'currency_field': 'currency_id'}"
-                />
-            </field>
             <field name="purchase_method" position="attributes">
                 <attribute name="invisible">1</attribute>
             </field>
@@ -96,36 +89,6 @@
         </field>
     </record>
 
-    <record id="product_template_edit_price_tree_view" model="ir.ui.view">
-        <field name="name">product.template.edit.price.tree</field>
-        <field name="model">product.template</field>
-        <field name="priority" eval="100" />
-        <field name="arch" type="xml">
-            <tree string="Product" editable="top" edit="true">
-                <field name="name" readonly="1" />
-                <field
-                    name="main_seller_id"
-                    string="Main Seller"
-                    readonly="1"
-                    options="{'no_open': True, 'no_create': True, 'no_create_edit': True }"
-                />
-                <field name="purchase_price" />
-                <field
-                    name="uom_po_id"
-                    readonly="1"
-                    options="{'no_open': True, 'no_create': True, 'no_create_edit':True }"
-                />
-                <field name="suggested_price" readonly="1" />
-                <field name="list_price" readonly="1" />
-                <field
-                    name="uom_id"
-                    readonly="1"
-                    options="{'no_open': True, 'no_create': True, 'no_create_edit':True }"
-                />
-            </tree>
-        </field>
-    </record>
-
     <record model="ir.ui.view" id="beesdoo_product_label_form">
         <field name="name">bees.product.label.form</field>
         <field name="model">beesdoo.product.label</field>
@@ -180,41 +143,6 @@
         </field>
     </record>
 
-    <record model="ir.ui.view" id="beesdoo_product_category_list">
-        <field name="name">beesdoo.product.category.list</field>
-        <field name="model">product.category</field>
-        <field name="inherit_id" ref="product.product_category_list_view" />
-        <field name="arch" type="xml">
-            <field name="display_name" position="after">
-                <field name="profit_margin" />
-                <field name="should_round_suggested_price" />
-            </field>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="beesdoo_product_category_form">
-        <field name="name">beesdoo.product.category.form</field>
-        <field name="model">product.category</field>
-        <field name="inherit_id" ref="product.product_category_form_view" />
-        <field name="arch" type="xml">
-            <field name="parent_id" position="after">
-                <field name="profit_margin" />
-                <field name="should_round_suggested_price" />
-            </field>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="beesdoo_product_supplierinfo_tree_view">
-        <field name="name">beesdoo.product.supplierinfo.tree</field>
-        <field name="model">product.supplierinfo</field>
-        <field name="inherit_id" ref="product.product_supplierinfo_tree_view" />
-        <field name="arch" type="xml">
-            <field name="price" position="replace">
-                <field name="price" />
-            </field>
-        </field>
-    </record>
-
     <record model="ir.actions.act_window" id="action_labels">
         <field name="name">Product Labels</field>
         <field name="res_model">beesdoo.product.label</field>
@@ -251,17 +179,6 @@
             <field name="name" position="after">
                 <field name="type" />
             </field>
-        </field>
-    </record>
-
-    <record id="beesdoo_product_res_parter_form" model="ir.ui.view">
-        <field name="name">res.partner.form</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="base.view_partner_form" />
-        <field name="arch" type="xml">
-            <group name="purchase" position="inside">
-                <field name="profit_margin" />
-            </group>
         </field>
     </record>
