+++ beesdoo_product/wizard/__init__.py
@@ -1,2 +1,1 @@
 from . import label_printing_utils
-from . import adapt_sales_price_wizard
