+++ sale_suggested_price/__init__.py
@@ -1,0 +1,2 @@
+from . import models
+from . import wizard
