+++ sale_suggested_price/models/__init__.py
@@ -1,0 +1,4 @@
+from . import product_category
+from . import product_template
+from . import res_config_settings
+from . import res_partner
