+++ sale_suggested_price/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Beescoop - Cellule IT
+* `Coop IT Easy SCRLfs <https://coopiteasy.be>`_:
