+++ sale_suggested_price/views/product_category_views.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record model="ir.ui.view" id="product_category_list_view">
+        <field name="name">sale_suggested_price.product.category.list</field>
+        <field name="model">product.category</field>
+        <field name="inherit_id" ref="product.product_category_list_view" />
+        <field name="arch" type="xml">
+            <field name="display_name" position="after">
+                <field name="profit_margin" />
+                <field name="should_round_suggested_price" />
+            </field>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="product_category_form_view">
+        <field name="name">sale_suggested_price.product.category.form</field>
+        <field name="model">product.category</field>
+        <field name="inherit_id" ref="product.product_category_form_view" />
+        <field name="arch" type="xml">
+            <field name="parent_id" position="after">
+                <field name="profit_margin" />
+                <field name="should_round_suggested_price" />
+            </field>
+        </field>
+    </record>
+</odoo>
