+++ sale_suggested_price/views/product_supplierinfo_views.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record model="ir.ui.view" id="product_supplierinfo_tree_view">
+        <field name="name">sale_suggested_price.product.supplierinfo.tree</field>
+        <field name="model">product.supplierinfo</field>
+        <field name="inherit_id" ref="product.product_supplierinfo_tree_view" />
+        <field name="arch" type="xml">
+            <field name="price" position="replace">
+                <field name="price" />
+            </field>
+        </field>
+    </record>
+</odoo>
