+++ sale_suggested_price/views/product_template_views.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record model="ir.ui.view" id="product_template_only_form_view">
+        <field name="name">sale_suggested_price.template.product.form</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="product.product_template_only_form_view" />
+        <field name="arch" type="xml">
+            <field name="list_price" position="after">
+                <field
+                    name="suggested_price"
+                    widget='monetary'
+                    options="{'currency_field': 'currency_id'}"
+                />
+            </field>
+        </field>
+    </record>
+    <record id="product_template_edit_price_tree_view" model="ir.ui.view">
+        <field name="name">sale_suggested_price.template.edit.price.tree</field>
+        <field name="model">product.template</field>
+        <field name="priority" eval="100" />
+        <field name="arch" type="xml">
+            <tree string="Product" editable="top" edit="true">
+                <field name="name" readonly="1" />
+                <!-- TODO: main_seller_id is not a part of this module -->
+                <!-- <field
+                    name="main_seller_id"
+                    string="Main Seller"
+                    readonly="1"
+                    options="{'no_open': True, 'no_create': True, 'no_create_edit': True }"
+                /> -->
+                <field name="purchase_price" />
+                <field
+                    name="uom_po_id"
+                    readonly="1"
+                    options="{'no_open': True, 'no_create': True, 'no_create_edit':True }"
+                />
+                <field name="suggested_price" readonly="1" />
+                <field name="list_price" readonly="1" />
+                <field
+                    name="uom_id"
+                    readonly="1"
+                    options="{'no_open': True, 'no_create': True, 'no_create_edit':True }"
+                />
+            </tree>
+        </field>
+    </record>
+</odoo>
