+++ sale_suggested_price/views/res_config_settings_views.xml
