+++ sale_suggested_price/wizard/__init__.py
@@ -1,0 +1,1 @@
+from . import adapt_sales_price_wizard
