+++ setup/sale_suggested_price/odoo/addons/sale_suggested_price
@@ -1,0 +1,1 @@
+../../../../sale_suggested_price
